re #24
Don't rely on self.defaultLoc, but generate a new one whenever we need it. Slightly more work, but then we know it is correct.

While working on this, the test files showed a change in the way axis maps handle the extremes.
It uses to be that if the extremes were not specifically mapped, they would map to themselves. So in the case of (Tests/test.py)
```a.map = [(500,250)]```
`(0,0)` and `(1000,1000)` were implied.
This does not seem to be the case anymore in fontTools. The extremes are no longer implied, and the default location maps to `-250`, and the test fails because there is no default font. It can be solved by explicitly adding the min and max values to the map.