WIP Refactoring
This refactors UFOprocessor from subclassing from to wrapping a fonttools designspace object in the hopes to have some buffer between the projects.

In this state ufoProcessor supports discrete axes: it will collect the right sources for the continuous bits and make glyphs, font info and kerning.

Todo: all the testing. Also the live-reloading thing is not part of this.