WIP Fix missing space.tight issue
So, if a source ufo has a variation on a space glyph or some other whitespace, like "space.tight", this will fail the checkGlyphIsEmpty test because it has no reognised whitespace unicode. This fix bluntly looks for "space" in the glyphname. While I'm aware this is not a comprehensive approach, it is low risk.