Add getDefaultFont
Check that indeed getDefaultFont and findDefaultFont find a default in a designspace with discrete axes. It will use the default value for each discete axis if no discreteLocation is given.

Updates to the README.md. Needs more examples and removal of old UFOProcessor stuff. (not just yet, it is a reminder of what I need to rewrite)