fixes clipThisLocation
clipThisLocation used the user-space coords to clip design-space locations. That wasn't good. So now it converts axis.minimum, axis.maximum to designspace values before clipping.

Values in discrete axes are not mapped before clipping - this is a question really. Currently assuming this will not happen.

Changed the name to clipDesignLocation to emphasise that we're working with designspace coords.

makeOneGlyph now expects a designspace location. Bend flag removed.

Fix typo on varModels.py