loadFonts to check if there is a logger when debug==True
For the following use case:
UFOProcessor object is initialised with `debug=False`, no logger is started.
Then, in a scripted context, the debug flag is set to True. But this is just an attribute, so there is no logger.
Then `loadFonts` checks both `debug` and `logger` and calls `startLog` to create one if needed.